Show progress notifications in classify one
