+++ digits/inference/images/classification/__init__.py
@@ -2,5 +2,6 @@
 from __future__ import absolute_import
 
 from .job import ImageInferenceClassificationJob
+from .job import ImageInferenceClassifyOneJob
 from .job import ImageInferenceClassifyManyJob
 from .job import ImageInferenceTopNJob
